# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clash Ren FF Tournament Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments:
    get:
      operationId: listTournaments
      tags: [tournaments]
      summary: List tournaments
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [upcoming, ongoing, completed]
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournament"
...
  /tournaments/{id}:
    get:
      operationId: getTournament
      tags: [tournaments]
      summary: Get tournament by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tournaments/{id}/join:
    post:
      operationId: joinTournament
      tags: [tournaments]
      summary: Join a tournament (deducts diamond entry fee)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successfully joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Cannot join (insufficient diamonds, already joined, full)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Tournament:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        gameMode:
          type: string
        entryFeeDiamonds:
          type: integer
        prizePoolDiamonds:
          type: integer
        maxSlots:
          type: integer
        filledSlots:
          type: integer
        startTime:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        roomId:
          type: string
          nullable: true
        roomPassword:
          type: string
          nullable: true
        isJoined:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - gameMode
        - entryFeeDiamonds
        - prizePoolDiamonds
        - maxSlots
        - filledSlots
        - startTime
        - status
        - isJoined
        - createdAt
...